feat: add ability to deploy built nestjs app to lambda.
