+++ cdk.ts
@@ -8,14 +8,8 @@ const stack = new cdk.Stack(app, 'CartServiceStack', {
   env: { region: process.env.AWS_REGION as string },
 });
 
-const cartServiceLambdaLayer = new lambda.LayerVersion(stack, 'cartServiceLayer', {
-  code: lambda.Code.fromAsset('node_modules'),
-  compatibleRuntimes: [lambda.Runtime.NODEJS_18_X],
-});
-
 new NodejsFunction(stack, 'cartServiceLambda', {
   runtime: lambda.Runtime.NODEJS_18_X,
-  entry: 'dist/src/handler.js',
+  entry: 'dist/main.js',
   handler: 'handler',
-  layers: [cartServiceLambdaLayer],
 });
