--- src/handler.ts
@@ -1,25 +1,0 @@
-import { createServer, proxy, Response } from '@vendia/serverless-express';
-import express from 'express';
-import { Server } from 'http';
-
-import { createApp } from './main';
-
-let cachedServer: Server;
-
-async function bootstrap(): Promise<Server> {
-  const expressApp = express();
-
-  const app = await createApp(expressApp);
-  await app.init();
-
-  return createServer(expressApp);
-}
-
-export async function handler(event: any, context): Promise<Response> {
-  if (!cachedServer) {
-    const server = await bootstrap();
-    cachedServer = server;
-  }
-
-  return proxy(cachedServer, event, context, 'PROMISE').promise;
-}
